refactor: import 개선 및 불필요함수 제거
