--- src/advanced/src/order/components/ShippingSection.tsx
@@ -1,8 +1,0 @@
-export function ShippingNotice() {
-  return (
-    <div className="flex justify-between text-sm tracking-wide text-gray-400">
-      <span>Shipping</span>
-      <span>Free</span>
-    </div>
-  );
-}
