+++ src/advanced/src/product/utils/getStockStatus.ts
