+++ src/advanced/src/product/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './stockUtils';
+export * from './getStockStatus';
 export * from './productStatus';
